Address feedback on moving util to tests
