+++ tests/helper_functions.py
@@ -11,6 +11,7 @@ import pandas as pd
 
 import mlflow.pyfunc.scoring_server as pyfunc_scoring_server
 import mlflow.pyfunc
+from mlflow.utils.file_utils import read_yaml, write_yaml
 
 
 def random_int(lo=1, hi=1e10):
@@ -137,3 +138,18 @@ def _evaluate_scoring_proc(proc, port, data, content_type, activity_polling_time
         print("-------------------------STDOUT------------------------------")
         print(proc.stdout.read())
         print("==============================================================")
+
+
+class safe_edit_yaml(object):
+    def __init__(self, root, file_name, edit_func):
+        self._root = root
+        self._file_name = file_name
+        self._edit_func = edit_func
+        self._original = read_yaml(root, file_name)
+
+    def __enter__(self):
+        new_dict = self._edit_func(self._original.copy())
+        write_yaml(self._root, self._file_name, new_dict, overwrite=True)
+
+    def __exit__(self, *args):
+        write_yaml(self._root, self._file_name, self._original, overwrite=True)
