+++ tests/utils/test_file_utils.py
@@ -13,7 +13,7 @@ from mlflow.utils import file_utils
 from mlflow.utils.file_utils import get_parent_dir, _copy_file_or_tree, TempDir
 from tests.projects.utils import TEST_PROJECT_DIR
 
-from tests.helper_functions import random_int, random_file
+from tests.helper_functions import random_int, random_file, safe_edit_yaml
 
 
 def test_yaml_read_and_write(tmpdir):
@@ -38,7 +38,7 @@ def test_yaml_read_and_write(tmpdir):
         return old_dict
 
     assert "more_text" not in file_utils.read_yaml(temp_dir, yaml_file)
-    with file_utils.safe_edit_yaml(temp_dir, yaml_file, edit_func):
+    with safe_edit_yaml(temp_dir, yaml_file, edit_func):
         editted_dict = file_utils.read_yaml(temp_dir, yaml_file)
         assert "more_text" in editted_dict
         assert editted_dict["more_text"] == u"西班牙语"
